cutted the request for data made by the client into an object
this is referenced by both the game.js and the manage.js, rather than
being defined locally the same way in both the two files. Also fixed
a bug in the html builder for the container handout which had not yet
heard that the property name has been renamed to label.